backend/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
boto3==1.28.3
botocore==1.31.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.1.5
django-allauth==0.52.0
django-apscheduler==0.6.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-leaflet==0.28.3
django-mptt==0.14.0
django-ordered-model==3.7.4
django-parler==2.3
django-polymorphic==3.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.4.0
prettyconf==2.2.1
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gettext==4.1
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
whitenoise==6.3.0
